docs/requirements.txt
sphinx>=7.2
pydata-sphinx-theme>=0.14
numpydoc>=1.6
sphinx-copybutton>=0.5
sphinx-design>=0.5
myst-parser>=2.0
sphinx-autodoc-typehints>=1.24
sphinxcontrib-mermaid>=0.9
